filterx-jit: move to separate directory
This is a preparation step for introducing another FilterXJIT-related
unit (FFI).
