+++ lib/filterx/CMakeLists.txt
@@ -1,4 +1,7 @@
+add_subdirectory(jit)
+
 set(FILTERX_HEADERS
+    ${FILTERX_JIT_HEADERS}
     filterx/expr-arithmetic-operators.h
     filterx/expr-assign.h
     filterx/expr-boolalg.h
@@ -88,6 +91,7 @@ set(FILTERX_HEADERS
     )
 
 set(FILTERX_SOURCES
+    ${FILTERX_JIT_SOURCES}
     filterx/expr-arithmetic-operators.c
     filterx/expr-assign.c
     filterx/expr-boolalg.c
