+++ lib/filterx/jit/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(FILTERX_JIT_HEADERS
+    jit/jit.h
+    PARENT_SCOPE
+    )
+
+set(FILTERX_JIT_SOURCES
+    jit/jit.c
+    PARENT_SCOPE
+    )
