+++ lib/filterx/jit/Makefile.am
@@ -1,0 +1,9 @@
+filterxjitincludedir = ${pkgincludedir}/filterx/jit
+
+EXTRA_DIST += lib/filterx/jit/CMakeLists.txt
+
+filterxjitinclude_HEADERS = \
+	lib/filterx/jit/jit.h
+
+filterxjit_sources = \
+	lib/filterx/jit/jit.c
